+++ tests/proxmoxsandboxtest/test_proxmox_sandbox_environment_e2e.py
@@ -3,7 +3,6 @@ from typing import Dict
 
 import pytest
 from inspect_ai.util import SandboxConnection, SandboxEnvironment
-from pytest import raises
 
 from proxmoxsandbox._impl.qemu_commands import QemuCommands
 from proxmoxsandbox._impl.sdn_commands import SdnCommands
@@ -314,31 +313,6 @@ async def test_ova() -> None:
             environments=envs_dict,
             interrupted=False,
         )
-
-
-async def test_at_least_one_sandbox() -> None:
-    sandbox_env_config = ProxmoxSandboxEnvironmentConfig(
-        vms_config=(
-            VmConfig(
-                vm_source_config=VmSourceConfig(
-                    built_in="ubuntu24.04",
-                ),
-                name="non-sandbox-1",
-                is_sandbox=False,
-            ),
-            VmConfig(
-                vm_source_config=VmSourceConfig(
-                    built_in="ubuntu24.04",
-                ),
-                name="non-sandbox-2",
-                is_sandbox=False,
-            ),
-        )
-    )
-    with raises(ValueError) as e_info:
-        await setup_sandbox("ta1s", sandbox_env_config)
-    assert "No default sandbox found" in str(e_info.value)
-    await ProxmoxSandboxEnvironment.cli_cleanup(id=None)
 
 
 async def test_connect(proxmox_sandbox_environment: ProxmoxSandboxEnvironment) -> None: